+++ crates/ptool-lua/src/lua_api.rs
@@ -75,6 +75,7 @@ fn create_ptool_module(
     let datetime_module = create_ptool_datetime_module(lua, Rc::clone(&world))?;
     let db_module = create_ptool_db_module(lua, Rc::clone(&world))?;
     let redis_module = create_ptool_redis_module(lua, Rc::clone(&world))?;
+    let consul_module = create_ptool_consul_module(lua, Rc::clone(&world))?;
     let s3_module = create_ptool_s3_module(lua, Rc::clone(&world))?;
     let git_module = create_ptool_git_module(lua, Rc::clone(&world))?;
     let ssh_module = create_ptool_ssh_module(lua, Rc::clone(&world))?;
@@ -117,6 +118,7 @@ fn create_ptool_module(
     module.set("datetime", datetime_module)?;
     module.set("db", db_module)?;
     module.set("redis", redis_module)?;
+    module.set("consul", consul_module)?;
     module.set("s3", s3_module)?;
     module.set("git", git_module)?;
     module.set("ssh", ssh_module)?;
@@ -415,6 +417,17 @@ fn create_ptool_redis_module(
         lua.create_function(move |_, value: Value| world.borrow().redis_connect(value))?;
     redis_module.set("connect", connect_fn)?;
     Ok(redis_module)
+}
+
+fn create_ptool_consul_module(
+    lua: &Lua,
+    world: Rc<RefCell<crate::LuaWorld>>,
+) -> mlua::Result<Table> {
+    let consul_module = lua.create_table()?;
+    let connect_fn =
+        lua.create_function(move |_, options: Table| world.borrow().consul_connect(options))?;
+    consul_module.set("connect", connect_fn)?;
+    Ok(consul_module)
 }
 
 fn create_ptool_s3_module(lua: &Lua, world: Rc<RefCell<crate::LuaWorld>>) -> mlua::Result<Table> {